implement deleting by id not found case

+++ src/main/java/guru/springframework/spring6restmvc/service/BeerServiceJpa.java
@@ -50,8 +50,12 @@ public class BeerServiceJpa implements BeerService {
     }
 
     @Override
-    public void deleteBeerById(UUID id) {
-        beerRepository.deleteById(id);
+    public boolean deleteBeerById(UUID id) {
+        if (beerRepository.existsById(id)) {
+            beerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
